Fix saves of user display name
User display names would not survive a restart of the service properly.
Instead of trying to key them off of the express_sid, we're now storying
the finalized display name (cn from LDAP) as part of the user object.

When used in conjunction with ep_disable_change_author_name you can lock
in users to their display names (must set requireAuthenication for all
of this to really take hold)