Fix LDAP connections hanging around
Ran into a problem where LDAP connections were acreting in the system.
Particularly from monitoring of the system since no cookies were saved
for the monitoring system (on purpose). I ended up having to pretty much
completely re-implement the ldapauth plugin so now just use ldapjs
directly instead of super-classing it.

Bump of the version to 0.1.0 to be a little more accurate in it's
version.